+++ src/metatags/handler.js
@@ -13,12 +13,13 @@
 import RUMAPIClient from '@adobe/spacecat-shared-rum-api-client';
 import { Audit } from '@adobe/spacecat-shared-data-access';
 import { calculateCPCValue } from '../support/utils.js';
-import { getObjectFromKey } from '../utils/s3-utils.js';
+import { getObjectFromKey, getObjectKeysUsingPrefix } from '../utils/s3-utils.js';
 import SeoChecks from './seo-checks.js';
 import { AuditBuilder } from '../common/audit-builder.js';
 import { wwwUrlResolver } from '../common/index.js';
 import metatagsAutoSuggest from './metatags-auto-suggest.js';
 import { convertToOpportunity } from '../common/opportunity.js';
+import { getTopPagesForSiteId } from '../canonical/handler.js';
 import { getIssueRanking, getBaseUrl } from './opportunity-utils.js';
 import {
   DESCRIPTION,
@@ -89,7 +90,7 @@ export async function opportunityAndSuggestions(finalUrl, auditData, context) {
   log.info(`Successfully synced Opportunity And Suggestions for site: ${auditData.siteId} and ${auditType} audit type.`);
 }
 
-export async function fetchAndProcessPageObject(s3Client, bucketName, url, key, log) {
+export async function fetchAndProcessPageObject(s3Client, bucketName, key, prefix, log) {
   const object = await getObjectFromKey(s3Client, bucketName, key, log);
   if (!object?.scrapeResult?.tags || typeof object.scrapeResult.tags !== 'object') {
     log.error(`No Scraped tags found in S3 ${key} object`);
@@ -101,9 +102,12 @@ export async function fetchAndProcessPageObject(s3Client, bucketName, url, key, 
     return null;
   }
 
-  const pageUrl = object.finalUrl ? new URL(object.finalUrl).pathname
-    : new URL(url).pathname;
+  let pageUrl = object.finalUrl ? new URL(object.finalUrl).pathname
+    : key.slice(prefix.length - 1).replace('/scrape.json', ''); // Remove the prefix and scrape.json suffix
   // handling for homepage
+  if (pageUrl === '') {
+    pageUrl = '/';
+  }
   return {
     [pageUrl]: {
       title: object.scrapeResult.tags.title,
@@ -206,9 +210,12 @@ export async function metatagsAutoDetect(site, pagesSet, context) {
   const { log, s3Client } = context;
   // Fetch site's scraped content from S3
   const bucketName = context.env.S3_SCRAPER_BUCKET_NAME;
+  const prefix = `scrapes/${site.getId()}/`;
+  const scrapedObjectKeys = await getObjectKeysUsingPrefix(s3Client, bucketName, prefix, log);
   const extractedTags = {};
-  const pageMetadataResults = await Promise.all([...pagesSet]
-    .map(([url, path]) => fetchAndProcessPageObject(s3Client, bucketName, url, path, log)));
+  const pageMetadataResults = await Promise.all(scrapedObjectKeys
+    .filter((key) => pagesSet.has(key))
+    .map((key) => fetchAndProcessPageObject(s3Client, bucketName, key, prefix, log)));
   pageMetadataResults.forEach((pageMetadata) => {
     if (pageMetadata) {
       Object.assign(extractedTags, pageMetadata);
@@ -216,7 +223,7 @@ export async function metatagsAutoDetect(site, pagesSet, context) {
   });
   const extractedTagsCount = Object.entries(extractedTags).length;
   if (extractedTagsCount === 0) {
-    log.error(`Failed to extract tags from scraped content for bucket ${bucketName}`);
+    log.error(`Failed to extract tags from scraped content for bucket ${bucketName} and prefix ${prefix}`);
   }
 
   // Perform SEO checks
@@ -235,16 +242,38 @@ export async function metatagsAutoDetect(site, pagesSet, context) {
   };
 }
 
+/**
+ * Transforms a URL into a scrape.json path for a given site
+ * @param {string} url - The URL to transform
+ * @param {string} siteId - The site ID
+ * @returns {string} The path to the scrape.json file
+ */
+function getScrapeJsonPath(url, siteId) {
+  const pathname = new URL(url).pathname.replace(/\/$/, '');
+  return `scrapes/${siteId}${pathname}/scrape.json`;
+}
+
 export async function runAuditAndGenerateSuggestions(context) {
   const {
-    site, audit, finalUrl, log, scrapeResultPaths,
+    site, audit, finalUrl, log, dataAccess,
   } = context;
-  log.info(scrapeResultPaths);
+  // Get top pages for a site
+  const siteId = site.getId();
+  const topPages = await getTopPagesForSiteId(dataAccess, siteId, context, log);
+  const includedURLs = await site?.getConfig()?.getIncludedURLs('meta-tags') || [];
+
+  // Transform URLs into scrape.json paths and combine them into a Set
+  const topPagePaths = topPages.map((page) => getScrapeJsonPath(page.url, siteId));
+  const includedUrlPaths = includedURLs.map((url) => getScrapeJsonPath(url, siteId));
+  const totalPagesSet = new Set([...topPagePaths, ...includedUrlPaths]);
+
+  log.info(`Received topPages: ${topPagePaths.length}, includedURLs: ${includedUrlPaths.length}, totalPages to process after removing duplicates: ${totalPagesSet.size}`);
+
   const {
     seoChecks,
     detectedTags,
     extractedTags,
-  } = await metatagsAutoDetect(site, scrapeResultPaths, context);
+  } = await metatagsAutoDetect(site, totalPagesSet, context);
 
   // Calculate projected traffic lost
   const {
@@ -326,6 +355,6 @@ export async function submitForScraping(context) {
 export default new AuditBuilder()
   .withUrlResolver((site) => site.getBaseURL())
   .addStep('submit-for-import-top-pages', importTopPages, AUDIT_STEP_DESTINATIONS.IMPORT_WORKER)
-  .addStep('submit-for-scraping', submitForScraping, AUDIT_STEP_DESTINATIONS.SCRAPE_CLIENT)
+  .addStep('submit-for-scraping', submitForScraping, AUDIT_STEP_DESTINATIONS.CONTENT_SCRAPER)
   .addStep('run-audit-and-generate-suggestions', runAuditAndGenerateSuggestions)
   .build();